internal/cliconfig: Move CLI config package
