+++ internal/cliconfig/config.go
@@ -1,4 +1,4 @@
-package config
+package cliconfig
 
 import (
 	"encoding/json"
