+++ internal/cliconfig/errors.go
@@ -1,4 +1,4 @@
-package config
+package cliconfig
 
 import (
 	"fmt"
