+++ internal/cliconfig/file.go
@@ -1,4 +1,4 @@
-package config
+package cliconfig
 
 import (
 	"fmt"
