+++ internal/cliconfig/remote.go
@@ -1,4 +1,4 @@
-package config
+package cliconfig
 
 import (
 	"crypto/x509"
